qt: initial port to new libsigrok API
This does the scanning properly according to the new way, and saves the
discovered devices to a GSList stored in the MainWindow instance. The
rest of the code should use the MainWindow::getDevices() getter to
retrieve these, and use the data (which is struct sr_dev_inst *) to
call libsigrok API functions.